feat!(languages): Major update
This update has included types for many other languages and the base
command for converting has been changed to use a generic command.
The command now is changed from `ConvertJSONtoTS` and
`ConvertJSONtoTSBuffer` to `ConvertJSONtoLang` and
`ConvertJSONtoLangBuffer` and you can specify your lang which you want
in the end.
Please make the changes in your configurations and refer to the README
for more information.